% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builder.R
\name{Builder}
\alias{Builder}
\title{Build the actual word file for publication. This word file still needs to be compiled to a PDF before sending.}
\usage{
Builder(
  CP.issue = CP.issue,
  abstr.txt = abstr.txt,
  pm.extract = pm.extract,
  pm.table = pm.table,
  wd.data = wd.data
)
}
\arguments{
\item{CP.issue}{Variable should be in globalenv}

\item{abstr.txt}{Variable should be in globalenv}

\item{pm.extract}{Variable should be in globalenv}

\item{pm.table}{Variable should be in globalenv}
}
\description{
Build the actual word file for publication. This word file still needs to be compiled to a PDF before sending.
}
