% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor.R
\name{Extractor}
\alias{Extractor}
\title{Extracts the PMIDs from the word documents in the Dropbox folder}
\usage{
Extractor(wd.data, callout = T)
}
\arguments{
\item{wd.data}{Path to the folder in which the word documents, having the PMIDs, are.}
}
\value{
\itemize{
\item 
}
}
\description{
Extracts the PMIDs from the word documents in the Dropbox folder
}
\examples{
-

}
