% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Receiver.R
\name{Receiver}
\alias{Receiver}
\title{Retrieves the PMIDs from the word documents from the Dropbox of 4CP}
\usage{
Receiver(wd.data)
}
\arguments{
\item{wd.data}{Path to the folder in which the word documents, having the PMIDs, are.}
}
\value{
\itemize{
\item 
}
}
\description{
Retrieves the PMIDs from the word documents from the Dropbox of 4CP
}
\examples{
-

}
