% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Builder.R
\name{Builder}
\alias{Builder}
\title{Build the actual word file for publication. This word file still needs to be compiled to a PDF before sending.}
\usage{
Builder(wd.data)
}
\arguments{
\item{wd.data}{Path to data}
}
\description{
Build the actual word file for publication. This word file still needs to be compiled to a PDF before sending.
}
